+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/SampleParkAuto.java
@@ -13,7 +13,7 @@ import org.firstinspires.ftc.teamcode.subsystems.Arm;
 import org.firstinspires.ftc.teamcode.subsystems.Claw;
 import org.firstinspires.ftc.teamcode.subsystems.Imu;
 import org.firstinspires.ftc.teamcode.subsystems.Lift;
-import org.firstinspires.ftc.teamcode.vision.LimelightLocalization;
+import org.firstinspires.ftc.teamcode.visionex.LimelightLocalization;
 import org.firstinspires.ftc.teamcode.utils.DriverHubHelp;
 import org.firstinspires.ftc.teamcode.utils.GamepadEvents;
