+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/nearSideNetAuto.java
@@ -8,7 +8,7 @@ import org.firstinspires.ftc.teamcode.roadrunner.tuning.MecanumDrive;
 import org.firstinspires.ftc.teamcode.subsystems.Arm;
 import org.firstinspires.ftc.teamcode.subsystems.Imu;
 import org.firstinspires.ftc.teamcode.subsystems.Lift;
-import org.firstinspires.ftc.teamcode.vision.LimelightLocalization;
+import org.firstinspires.ftc.teamcode.visionex.LimelightLocalization;
 //import org.firstinspires.ftc.teamcode.roadrunner.tuning.ThreeDeadWheelLocalizer;
 import org.firstinspires.ftc.teamcode.utils.DriverHubHelp;
 import org.firstinspires.ftc.teamcode.utils.GamepadEvents;
