+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/vision/SampleDetection.java
@@ -7,6 +7,7 @@ import android.graphics.Rect;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.robotcore.internal.camera.calibration.CameraCalibration;
+import org.firstinspires.ftc.teamcode.visionex.Sample;
 import org.firstinspires.ftc.vision.VisionProcessor;
 import org.opencv.core.Mat;
 import org.openftc.easyopencv.OpenCvPipeline;
