+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/visionex/ResizingPipeline.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.vision;
+package org.firstinspires.ftc.teamcode.visionex;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.opencv.core.Mat;
