+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/visionex/Sample.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.vision;
+package org.firstinspires.ftc.teamcode.visionex;
 
     public enum Sample {
         RED,
