+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/visionex/ThresholdingPipeline.java
@@ -1,17 +1,12 @@
-package org.firstinspires.ftc.teamcode.vision;
+package org.firstinspires.ftc.teamcode.visionex;
 
 import org.opencv.core.Core;
 import org.opencv.core.Mat;
-import org.opencv.core.MatOfPoint;
 import org.opencv.core.Scalar;
 import org.opencv.imgproc.Imgproc;
 import org.openftc.easyopencv.OpenCvPipeline;
 
-import java.util.ArrayList;
-import java.util.List;
-
-public class ContourPipeline extends OpenCvPipeline {
-
+public class ThresholdingPipeline extends OpenCvPipeline {
     @Override
     public Mat processFrame(Mat input) {
         Scalar lower_blue = new Scalar(85,50, 40);
@@ -23,13 +18,6 @@ public class ContourPipeline extends OpenCvPipeline {
         Mat threshold = new Mat();
         Core.inRange(hsvMat, lower_blue, higher_blue, threshold);
 
-        List<MatOfPoint> contours = new ArrayList<>();
-        Mat hierachy = new Mat();
-
-        Imgproc.findContours(threshold, contours, hierachy,Imgproc.RETR_TREE, Imgproc.CHAIN_APPROX_SIMPLE);
-
-        Imgproc.drawContours(input, contours, -1, new Scalar(0, 255, 0), 3);
-
-        return null;
+        return threshold;
     }
 }
